tab fixes

+++ gnome-base/gdm/gdm-2.2.5.4-r5.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2002 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/gnome-base/gdm/gdm-2.2.5.4-r5.ebuild,v 1.11 2002/10/24 23:23:45 blizzy Exp $
+# $Header: /var/cvsroot/gentoo-x86/gnome-base/gdm/gdm-2.2.5.4-r5.ebuild,v 1.12 2002/11/02 10:12:09 azarah Exp $
 
 IUSE="nls gnome"
 
@@ -27,76 +27,80 @@ RDEPEND="${DEPEND}
 
 src_unpack() {
 
-	unpack "${A}"
+	unpack ${A}
 
-	cd "${S}/daemon"
+	cd ${S}/daemon
 	cp gdm.h gdm.h.orig
-	sed -e "s:/usr/bin/X11:/usr/X11R6/bin:g" gdm.h.orig > gdm.h
+	sed -e "s:/usr/bin/X11:/usr/X11R6/bin:g" \
+		gdm.h.orig > gdm.h
 	rm gdm.h.orig
 
-	cd "${S}/config"
+	cd ${S}/config
 	cp gdm.conf.in gdm.conf.in.orig
-	sed -e "s:/usr/bin/X11:/usr/X11R6/bin:g" gdm.conf.in.orig > gdm.conf.in
+	sed -e s:/usr/bin/X11:/usr/X11R6/bin:g \
+		gdm.conf.in.orig > gdm.conf.in
 	rm gdm.conf.in.orig
 }
 
 src_compile() {
-        local myconf
-        use nls || myconf="${myconf} --disable-nls"
-	./configure --host="${CHOST}" 					\
-		    --prefix=/usr					\
-		    --sysconfdir=/etc/X11				\
-		    --localstatedir=/var/lib				\
-		    --with-pam-prefix=/etc                              \
-                    ${myconf} || die
+	local myconf=""
+	use nls || myconf="${myconf} --disable-nls"
+	
+	./configure --host=${CHOST} \
+		--prefix=/usr \
+		--sysconfdir=/etc/X11 \
+		--localstatedir=/var/lib \
+		--with-pam-prefix=/etc \
+		${myconf} || die
 	emake || die
 }
 
 src_install() {
-        cd omf-install
-        cp Makefile Makefile.old
-        sed -e "s:scrollkeeper-update.*::g" Makefile.old > Makefile
-        rm Makefile.old
-        cd "${S}"
+	cd omf-install
+	cp Makefile Makefile.old
+	sed -e "s:scrollkeeper-update.*::g" Makefile.old > Makefile
+	rm Makefile.old
+	cd ${S}
 
-	make prefix="${D}/usr"						\
-	     sysconfdir="${D}/etc/X11"					\
-	     localstatedir="${D}/var/lib"				\
-	     PAM_PREFIX="${D}/etc"					\
-	     install || die
+	make prefix=${D}/usr \
+		sysconfdir=${D}/etc/X11 \
+		localstatedir=${D}/var/lib \
+		PAM_PREFIX=${D}/etc \
+		install || die
 
-	rm -f "${D}/etc/pam.d/gdm"
+	rm -f ${D}/etc/pam.d/gdm
 
 	# log
-	mkdir -p "${D}/var/lib/gdm"
-	chown gdm:gdm "${D}/var/lib/gdm"
-	chmod 750 "${D}/var/lib/gdm"
+	mkdir -p ${D}/var/lib/gdm
+	chown gdm.gdm ${D}/var/lib/gdm
+	chmod 750 ${D}/var/lib/gdm
   
 	# pam startup
 	dodir /etc/pam.d
 	insinto /etc/pam.d
-	doins "${FILESDIR}/${MY_V}/pam.d/gdm"
-	doins "${FILESDIR}/${MY_V}/pam.d/gdmconfig"
+	doins ${FILESDIR}/${MY_V}/pam.d/gdm
+	doins ${FILESDIR}/${MY_V}/pam.d/gdmconfig
 
 	# pam security
 	dodir /etc/security/console.apps
 	insinto /etc/security/console.apps
-	doins "${FILESDIR}/${MY_V}/security/console.apps/gdmconfig"
+	doins ${FILESDIR}/${MY_V}/security/console.apps/gdmconfig
 
 	# gnomerc
 	dodir /etc/X11/gdm
 	exeinto /etc/X11/gdm
-	doexe "${FILESDIR}/${MY_V}/gnomerc"
+	doexe ${FILESDIR}/${MY_V}/gnomerc
 
-	cd "${D}/etc/X11/gdm"
+	cd ${D}/etc/X11/gdm
 	for i in Init/Default PostSession/Default PreSession/Default gdm.conf
 	do
 		cp $i $i.orig
-		sed -e "s:/usr/bin/X11:/usr/X11R6/bin:g" $i.orig > $i
+		sed -e "s:/usr/bin/X11:/usr/X11R6/bin:g" \
+			$i.orig > $i
 		rm $i.orig
 	done
 
-	cd "${D}/etc/X11/gdm"
+	cd ${D}/etc/X11/gdm
 	cp gdm.conf gdm.conf.orig
 	
 	sed -e "s:0=/usr/X11R6/bin/X:0=/usr/X11R6/bin/X -dpi 100 -nolisten tcp dpms vt7:g" \
@@ -107,10 +111,10 @@ src_install() {
 
 	rm gdm.conf.orig
 
-	cd "${S}"
+	cd ${S}
 
 	#support for new session stuff
-	rm -rf "${D}/etc/X11/gdm/Sessions"
+	rm -rf ${D}/etc/X11/gdm/Sessions
 	dosym ../Sessions /etc/X11/gdm/Sessions
 	
 	dodoc ABOUT-NLS AUTHORS COPYING ChangeLog INSTALL NEWS README* TODO
@@ -118,7 +122,8 @@ src_install() {
 
 pkg_preinst() {
 	#support for new session stuff
-	if [ -d /etc/X11/gdm/Sessions ] ; then
+	if [ -d /etc/X11/gdm/Sessions -a ! -L ${ROOT}/etc/X11/gdm/Sessions ]
+	then
 		mv -f /etc/X11/gdm/Sessions /etc/X11/gdm/Sessions.old
 	fi
 }
@@ -165,8 +170,8 @@ pkg_postinst() {
 }
 
 pkg_postrm() {
-        echo ">>> Updating Scrollkeeper database..."
-        scrollkeeper-update >/dev/null 2>&1
+	echo ">>> Updating Scrollkeeper database..."
+	scrollkeeper-update >/dev/null 2>&1
 
 	echo
 	einfo "To remove GDM from startup please execute"
